requirements.txt
PyYAML>=6.0.1
requests>=2.31.0
qdrant-client>=1.7.0
streamlit>=1.29.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
celery>=5.3.0
redis>=5.0.0
langgraph>=0.1.0
pypdf>=4.0.0
python-docx>=1.1.0
